Backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pymongo>=4.3.3
python-dotenv>=1.0.0
python-multipart>=0.0.6

dnspython>=2.3.0

Pillow>=9.5.0

aiofiles>=23.1.0
email-validator>=2.0.0
pydantic>=1.10.7